requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.8.6
Flask-Cors==3.0.10
PyJWT==1.7.1